+++ packages/misskey-js/src/autogen/types.ts
@@ -8679,6 +8679,17 @@ export type operations = {
             /** @enum {string} */
             federation: 'all' | 'specified' | 'none';
             federationHosts: string[];
+            enableLlmTranslator: boolean;
+            enableLlmTranslatorRedisCache: boolean;
+            llmTranslatorRedisCacheTtl: number;
+            llmTranslatorBaseUrl: string | null;
+            llmTranslatorApiKey: string | null;
+            llmTranslatorModel: string | null;
+            llmTranslatorTemperature: number | null;
+            llmTranslatorTopP: number | null;
+            llmTranslatorMaxTokens: number | null;
+            llmTranslatorSysPrompt: string | null;
+            llmTranslatorUserPrompt: string | null;
           };
         };
       };
@@ -11024,6 +11035,17 @@ export type operations = {
           /** @enum {string} */
           federation?: 'all' | 'none' | 'specified';
           federationHosts?: string[];
+          enableLlmTranslator?: boolean;
+          enableLlmTranslatorRedisCache?: boolean;
+          llmTranslatorRedisCacheTtl?: number;
+          llmTranslatorBaseUrl?: string | null;
+          llmTranslatorApiKey?: string | null;
+          llmTranslatorModel?: string | null;
+          llmTranslatorTemperature?: number | null;
+          llmTranslatorTopP?: number | null;
+          llmTranslatorMaxTokens?: number | null;
+          llmTranslatorSysPrompt?: string | null;
+          llmTranslatorUserPrompt?: string | null;
         };
       };
     };